updates 'data' directory by reorganizing data components to be more moduler and renaming it to 'store'
