+++ src/app/components/navigation-bar/NavigationBar.component.jsx
@@ -7,6 +7,10 @@ class NavigationBar extends Component {
     constructor(props) {
         super(props);
 
+        this.state = {
+            
+        }
+
         this.initializeSideNav();
     }
