--- src/assets/data/experience/index.js
@@ -1,13 +1,0 @@
-import JSE_LexiaLearning from './experiences/JSE_LexiaLearning.js';
-import ICST_MIT from './experiences/ICST_MIT.js';
-import PAA_HPAC from './experiences/PAA_HPAC.js';
-import CSA_HarvardEdPortal from './experiences/CSA_HarvardEdPortal.js';
-
-const experience = [
-    JSE_LexiaLearning,
-    ICST_MIT,
-    PAA_HPAC,
-    CSA_HarvardEdPortal
-]
-
-export const EXPERIENCE_DATA = experience;
