--- src/assets/data/home/navigation-cards/index.js
@@ -1,3 +1,0 @@
-import navigationCards from './NavigationCards.js';
-
-export const NAVIGATION_CARDS_DATA = navigationCards
