--- src/assets/data/home/portrait-jumbotron/TypewriterPositions.js
@@ -1,8 +1,0 @@
-const typewriterPositions = [
-    "Front End Developer", 
-    "Problem Solver", 
-    "Junior Software Engineer", 
-    "Creative Thinker"
-];
-
-export default typewriterPositions;
