--- src/assets/data/home/portrait-jumbotron/index.js
@@ -1,10 +1,0 @@
-import typewriterPositions from './TypewriterPositions.js';
-import portrait from './../../../images/portrait_placeholder.png';
-
-const portraitJumbotron = {
-    typewriterPositions: typewriterPositions,
-    portrait: portrait,
-    name: 'Nazmuz Shakib Pranto'
-}
-
-export const PORTRAIT_JUMBOTRON_DATA = portraitJumbotron;
