--- src/assets/data/socials-footer/index.js
@@ -1,3 +1,0 @@
-import socialsFooter from './SocialsFooter.js';
-
-export const SOCIALS_FOOTER_DATA = socialsFooter;
