+++ src/assets/store/App.store.js
@@ -1,0 +1,17 @@
+import portfolioAuthorPortrait from './../images/portrait_placeholder.png';
+
+import {SOCIALS} from './aisles/Socials.aisle.js';
+import {NAVIGATION_ROUTES} from './aisles/NavigationRoutes.aisle.js';
+import {TYPEWRITER_POSITIONS} from './aisles/TypewriterPositions.aisle.js';
+import {EXPERIENCES} from './aisles/Experiences.aisle.js';
+
+const AppStore = {
+    portfolioAuthorName: 'Nazmuz Shakib Pranto',
+    portfolioAuthorPortrait: portfolioAuthorPortrait,
+    socials: SOCIALS,
+    navigationRoutes: NAVIGATION_ROUTES,
+    typewriterPositions: TYPEWRITER_POSITIONS,
+    experiences: EXPERIENCES
+}
+
+export default AppStore;
