+++ src/assets/store/aisles/Experiences.aisle.js
@@ -1,0 +1,13 @@
+import JSE_LexiaLearning from './../units/experiences/JSE_LexiaLearning.js';
+import ICST_MIT from './../units/experiences/ICST_MIT.js';
+import PAA_HPAC from './../units/experiences/PAA_HPAC.js';
+import CSA_HarvardEdPortal from './../units/experiences/CSA_HarvardEdPortal.js';
+
+const experiences = [
+    JSE_LexiaLearning,
+    ICST_MIT,
+    PAA_HPAC,
+    CSA_HarvardEdPortal
+]
+
+export const EXPERIENCES = experiences;
