+++ src/assets/store/aisles/TypewriterPositions.aisle.js
@@ -1,0 +1,8 @@
+const typewriterPositions = [
+    "Front End Developer",
+    "Problem Solver",
+    "Junior Software Engineer",
+    "Creative Thinker"
+];
+
+export const TYPEWRITER_POSITIONS = typewriterPositions;
