+++ src/assets/store/index.js
@@ -1,0 +1,3 @@
+import appStore from './App.store.js';
+
+export const APP_STORE = appStore;
